UTs on 64bit OS
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4612 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ quex/engine/loop/TEST/helper.py
@@ -464,7 +464,7 @@ show_next_character(QUEX_TYPE_ANALYZER* me)
             printf("next letter: <0x%02X>", (int)(buffer->_read_p[0]));
 
 #       if $$SHOW_POSITION$$
-        printf(" column_n: %i", me->counter._column_number_at_end);
+        printf(" column_n: %i", (int)me->counter._column_number_at_end);
 #       endif
         printf("\\n");
     }
